Refactored the text truncation function
Text truncation was very inaccurate under certain circumstances. The
position that it truncated to would vary greatly, sometimes cutting off
way too early, leaving up to half of the available space blank.

This resolves that by using a bitwise operator to adjust to the desired
length efficiently. It calculates the middle position for each loop,
and adjusts based on that.

It should be efficient, and accurate. Unfortunately due to some quirks
in the rendering process, I had to subtract 1 em (“M”) from the
available width to prevent certain characters from overflowing. This
results in a little bit less space used than is available, but the end
result is much better than the previous version.