Fix font registration for new version of PDFKit
FontKit, through PDFKit, takes the family name that we pass in
and uses it as a postscript name to try to find a font
variation, rather than registering our new font.  This check
throws an error since we do not have the necessary tables for
such a variation.  Not passing in the postscriptname allows the
font to be registered normally.