Mirror PDFKit `widthOfString()` functionality
The wrapping functionality within PDFKit checks the width of each
word individually, rather than checking the entire string upfront
(as Fluent Reports is doing). This results in slight size
discrepancies, as spaces are different sizes depending on the
following character.  This can lead to Fluent Reports'
`truncateText()` failing to remove as many characters as needed,
causing the band to wrap to the next line when it should cutoff.